+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to this project will be documented in this file.
 **ATTN**: This project uses [semantic versioning](http://semver.org/).
 
 ## [Unreleased]
+
+## [v1.3.0] - 2020-12-02
 ### Fixed
 - Fixed wrong number of bytes written in Packet WriteTo function.
 
@@ -65,7 +67,8 @@ changed.
 ### Added
 - Initial implementation.
 
-[Unreleased]: https://github.com/gorcon/rcon/compare/v1.2.4...HEAD
+[Unreleased]: https://github.com/gorcon/rcon/compare/v1.3.0...HEAD
+[v1.3.0]: https://github.com/gorcon/rcon/compare/v1.2.4...v1.3.0
 [v1.2.4]: https://github.com/gorcon/rcon/compare/v1.2.3...v1.2.4
 [v1.2.3]: https://github.com/gorcon/rcon/compare/v1.2.2...v1.2.3
 [v1.2.2]: https://github.com/gorcon/rcon/compare/v1.2.1...v1.2.2